\name{IndCompuestos-package}
\alias{IndCompuestos-package}
\alias{IndCompuestos}
\docType{package}
\title{Analisis de Decision Multicriterio}
\description{Este paquete contiene funciones para el calculo de indicadores compuestos, mediante diversos metodos de normalizacion y agregacion. 
Ademas contiene funciones para la obtencion de submatrices de subindicadores aleatorias y para el calculo de la medida de Shannon-Spearman.}
\details{
\tabular{ll}{
Package: \tab IndCompuestos\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-16\cr
License: \tab GPL-2\cr
}
El principal objetivo del paquete es proporcionar herramientas para el calculo de indicadores compuestos dentro del marco del Analisis de Decision Multicriterio.
}
\author{Minolli Santiago.
Maintainer: Minolli Santiago <minollisantiago@gmail.com>}
\references{}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
